정적 분석·ASan 을 쓸 수 있게 고치고, 그렇게 드러난 결함을 잡는다
ASan 과 clang-tidy 를 돌려 보니 둘 다 쓸 수 없는 상태였다. 도구를 고치는 것이 조사의 절반이었다.

Windows ASan 빌드는 한 번도 완주한 적이 없었다(CI 는 Linux ASan 만 돌린다). 세 군데가 독립적으로
막고 있었다:

1. ASan 런타임 DLL 을 Bin/ 에만 복사했다. 코드젠 도구는 BuildTools/ 에 있어 못 찾고 0xc0000135 로
   뜨지 못한다 — 빌드가 코드젠 단계에서 죽었다. 두 곳 모두에 복사한다.
2. /MD 를 FORCE 하는데(clang-cl ASan 은 디버그 CRT 를 지원하지 않는다) vcpkg 는 Debug 구성에서
   /MDd 라이브러리를 물려줘 _ITERATOR_DEBUG_LEVEL 이 어긋났다. 임포트 구성을 Release 로 매핑하고
   런타임 DLL 도 릴리스 쪽을 직접 staging 한다(applocal 경로는 툴체인에 박혀 있어, 그냥 두면
   z.dll 이 없어 CoreTest.exe 가 뜨지 못한다).
3. MSVC STL 이 ASan 주석 표시를 심어 비-ASan vcpkg 라이브러리와 또 어긋난다. 개별 매크로를 막으면
   다음 것이 나오므로(string -> vector -> optional) 우산 매크로 _DISABLE_STL_ANNOTATION 로 끈다.

그 위에 실제로 쓸 수 있게:

- 테스트 타임아웃을 ASan 에서 10배로. 평시 기준을 그대로 써서 5개가 전부 타임아웃으로 실패하고
  있었다 — 결함처럼 보이지만 설정 문제였다.
- CTest 환경에 detect_odr_violation=1. ODR 보고가 전부 SDK·CRT 헤더 전역이다(d3d11.h 의
  D3D11_DEFAULT, CRT 내부 _Avx2WmemEnabledWeakValue). 플러그인 DLL 이 여럿인 구조에서 영구 오탐이다.
  끄지(0) 않고 1 로 둬서 크기가 다른 진짜 ODR 버그는 계속 잡는다.
- SmokeTest 의 모듈 해제 후 적재 케이스는 ASan 에서 스킵한다(8/8 결정적 실패, 원인 미규명 —
  배제한 가설들과 재현 절차는 백로그 1-3).

clang-tidy 는 오탐이 신호를 덮고 있었다:

- .clang-tidy — 이 코드베이스에서 쓸 수 없다고 실측한 두 검사만 끈다(easily-swappable-parameters
  158건, EnumCastOutOfRange 39건 = 비트 플래그 조합). 끈 이유를 파일에 적었다.
- Scripts/lint/RunClangTidy.py — 한 명령으로 같은 설정. 핵심은 /Y- 로 MSVC PCH 를 무효화한 것이다.
  clang-tidy 는 그 PCH 를 쓸 수 없는데 /Yu·/FI 가 남아 헤더를 두 경로 표기로 두 번 파싱하고,
  #pragma once 가 듣지 않아 "redefinition of ..." 오류가 쏟아졌다 — 정의는 하나뿐인데도.
- 노이즈 포함 ~300건 -> 고유 110건. 종류별 판정은 백로그 1-2 에 적었다.

그렇게 해서 찾은 실제 결함:

- BattleState::applyMove 널 역참조. 종족·카탈로그 조회가 실패하면 pMove 가 nullptr 인데
  dmg == 0 분기가 pMove->_name 을 무방비로 읽었다. dmg > 0 분기는 널 검사를 함의해 안전했으므로
  정상 데이터로는 절대 걸리지 않는다 — 기술 데이터가 빠지면 크래시다.
- XML 역직렬화가 길이를 버렸다. initXmlDeserialization 이 const utf8* 를 받아 호출부가
  string_view::data() 를 넘기며 길이를 잃었다. 뷰가 더 큰 버퍼의 일부면 널 종단이 없어 파서가 끝을
  넘어 읽는다. XmlDocument::parse 는 원래부터 string_view + load_buffer(data, size) 로 안전하므로
  이 중간 계층만 구멍이었다. 인터페이스를 string_view 로 바꿔 길이를 끝까지 흘려보낸다.
- Windows ASan 에서 누수 검사가 도는 척만 했다. Linux/macOS 는 ASan 이면 LSAN 으로 갈라 두었는데
  Windows 만 ASan 여부를 보지 않고 CRT 검사를 골랐다. ASan 이 힙을 대체하므로 _Crt* 가 전부 무효다.
  유일한 증상이 "set but not used" 경고 둘이었다.
- int 곱셈 후 size_t 확대 네 곳(TileMap::indexOf, TileMap/TileMapXml 의 _width*_height,
  Defines.h·hashed_string.h 의 크기 상수). 넘친 뒤 확대되므로 캐스트가 값을 지켜 주는 것처럼
  보이지만 이미 틀린 값이다.
- FrameDoubleBuffer 의 기본 용량이 숫자로 박혀 있었다 -> 같은 파일이 이미 쓰는 상수로.

의도된 패턴 5곳은 NOLINTNEXTLINE + 이유 주석으로 남겼다(검사는 새 코드에 계속 살아 있게).

검증: Debug·Shipping 경고 0, lint 6/6, nogpu 5/5 양쪽, EngineTest 423/423, 네 백엔드 × 에디터
[Error] 0건, 패널 덤프 14창/빈 0개. ASan 하에서 CoreTest 167/167 · ReflectionTest 100 ·
EditorTest 34/34 가 보고 0건.

+++ Source/Engine/Utility/Xml/TileMapXml.cpp
@@ -63,7 +63,7 @@ namespace sw
             _height = 8;
         }
 
-        const size_t count = static_cast<size_t>( _width * _height );
+        const size_t count = static_cast<size_t>( _width ) * static_cast<size_t>( _height );
         _listWalkable.assign( count, 1 );
         _listEncounter.assign( count, 0 );
         _listPassThrough.assign( count, 0 );
@@ -193,7 +193,7 @@ namespace sw
         spawn.appendAttr( "y", _spawnY );
 
         XmlNode      tiles = root.appendChild( "tiles" );
-        const size_t count = static_cast<size_t>( _width * _height );
+        const size_t count = static_cast<size_t>( _width ) * static_cast<size_t>( _height );
         for ( size_t tileIndex = 0; tileIndex < count; ++tileIndex )
         {
             XmlNode       tileNode   = tiles.appendChild( "t" );
